pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "en"
version = "0.1.0"
description = "EN coordination for AI agent teams"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
